+++ packages/number-input/src/ErrorIcon/index.tsx
@@ -1,0 +1,1 @@
+export { ErrorIcon } from './ErrorIcon';
